+++ src/main/java/lee/code/essentials/commands/cmds/WorthCMD.java
@@ -29,7 +29,7 @@ public class WorthCMD implements CommandExecutor {
                     }
                 }
                 if (ItemSellValues.valueOf(name).getItem().equals(itemHand)) {
-                    int value = ItemSellValues.valueOf(name).getValue();
+                    long value = ItemSellValues.valueOf(name).getValue();
                     player.sendMessage(Lang.PREFIX.getString(null) + Lang.COMMAND_WORTH_SUCCESSFUL.getString(new String[] { plugin.getPU().formatMaterial(name), plugin.getPU().formatAmount(value) }));
                 } else player.sendMessage(Lang.PREFIX.getString(null) + Lang.ERROR_SELL_NOT_SELLABLE.getString(null));
             } else player.sendMessage(Lang.PREFIX.getString(null) + Lang.ERROR_SELL_NOT_SELLABLE.getString(null));
